Turning this into a suite of software called clipsite.
Clipsite will allow users to push their kindle clippings to the web so others will iknow what they are reading and what they find interesting about it.

My thoughts are that when the kindle is connected to a computer, we'll post the clippings to a web site.    That site will allow republishing of the clippings to facebook or blogs, etc via widgets and rss.

I read somewhere that individual books have clipping limits baked into the drm.   This site could push back to the kindle your latest clippings per book, allowing you to clip beyond the limit.  You'd just have to adhere to the limit on the kindle itself.

Bringing in the OSX files I use to grab the clippings any time the Kindle is mounted.
Also, the django models appear to be working.
The cmdline parser is doing a good job and the loader assumes user_id =1 for the owner of all the clippings.